Make component for info panel
